requirements.txt
allure-pytest
allure-python-commons
pytest
requests
selenium
webdriver-manager